Revised GCC:GOCART2G connection
Directly linked GOCART2G species are now friendly to nothing.
They are accessed via a field bundle, fSPC, containing GOCART2G
fields that are FriendlyTo GEOSCHEMCHEM.
The pass back:forth is simple & hardcoded
The internal mixture process is still the same. Added NITs and SO4s
to the IM fields for SS as if mixed with SALC.
MSL